+++ pcs_test/tier1/legacy/test_resource.py
@@ -3086,7 +3086,14 @@ monitor interval=20 (A-monitor-interval-20)
             self.temp_cib.name,
             "resource move-with-constraint DGroup rh7-1".split(),
         )
-        ac(o, "")
+        ac(
+            o,
+            (
+                "Warning: A move constraint has been created and the resource "
+                "'DGroup' may or may not move depending on other configuration"
+                "\n"
+            ),
+        )
         assert r == 0
 
         o, r = pcs(self.temp_cib.name, ["constraint"])